feat:생년월일입력 후 mbti 페이지 이동
